Changes from PR feedback suggestions in #3
    * Moved top-level files into structured folders:
        * Moved core logic to Retrograde/Core/
        * Moved infrastructure classes to Retrograde/Infra/
        * Moved UI classes to Retrograde/UI/
        * Moved Program.cs to Retrograde/ for consistency

    * Updated namespaces to reflect new folder structure:
        * Dn500BD.Retrograde.Core
        * Dn500BD.Retrograde.Infra
        * Dn500BD.Retrograde.UI

    * Normalized line endings in .editorconfig:
        * Changed end_of_line from crlf to lf

    * Removed stale files that were replaced by new structure:
        * Deleted old Program.cs and Retrograde/*.cs files

    * Added Utilities section to README.md:
        * Documented usage of audit.ps1 script
        * Removed unnecessary global.json explanation per team discussion

    * Dismissed several suggestions from PR as not applicable:
        * Reconnection logic for serial port rejected as unnecessary
        * Custom exception for SerialPortService deemed overengineering
        * Kept naming rule identifier verbose but functional
        * Declined to make ReadWithTimeoutAsync virtual without concrete need
