+++ .editorconfig
@@ -53,7 +53,7 @@ dotnet_naming_style.pascal_case.capitalization = pascal_case
 indent_style = space
 indent_size = 4
 tab_width = 4
-end_of_line = crlf
+end_of_line = lf
 charset = utf-8
 insert_final_newline = true
 trim_trailing_whitespace = true
