+++ Retrograde/Core/IDenonRemote.cs
@@ -1,4 +1,4 @@
-namespace Dn500BD.Retrograde;
+namespace Dn500BD.Retrograde.Core;
 
 public interface IDenonRemote
 {
